+++ src/textual/widgets/_footer.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
 from collections import defaultdict
+from typing import TYPE_CHECKING
 
 import rich.repr
 from rich.text import Text
@@ -10,6 +11,9 @@ from ..binding import Binding
 from ..containers import ScrollableContainer
 from ..reactive import reactive
 from ..widget import Widget
+
+if TYPE_CHECKING:
+    from ..screen import Screen
 
 
 @rich.repr.auto
@@ -128,8 +132,12 @@ class Footer(ScrollableContainer, can_focus=False, can_focus_children=False):
     """Convert 'ctrl+' prefix to '^'."""
     compact = reactive(False)
     """Display in compact style."""
+    _bindings_ready = reactive(False, repaint=False)
+    """True if the bindings are ready to be displayed."""
 
     def compose(self) -> ComposeResult:
+        if not self._bindings_ready:
+            return
         bindings = [
             (binding, enabled)
             for (_, binding, enabled) in self.screen.active_bindings.values()
@@ -157,9 +165,10 @@ class Footer(ScrollableContainer, can_focus=False, can_focus_children=False):
             )
 
     def on_mount(self) -> None:
-        def bindings_changed(screen) -> None:
-            if screen is self.screen:
-                self.call_next(self.recompose)
+        async def bindings_changed(screen: Screen) -> None:
+            self._bindings_ready = True
+            if self.is_attached and screen is self.screen:
+                await self.recompose()
 
         self.screen.bindings_updated_signal.subscribe(self, bindings_changed)