+++ tests/snapshot_tests/test_snapshots.py
@@ -1269,3 +1269,22 @@ def test_grid_gutter(snap_compare):
 def test_multi_keys(snap_compare):
     # https://github.com/Textualize/textual/issues/4542
     assert snap_compare(SNAPSHOT_APPS_DIR / "multi_keys.py")
+
+
+def test_data_table_in_tabs(snap_compare):
+    assert snap_compare(SNAPSHOT_APPS_DIR / "data_table_tabs.py")
+
+
+def test_auto_tab_active(snap_compare):
+    # https://github.com/Textualize/textual/issues/4593
+    assert snap_compare(SNAPSHOT_APPS_DIR / "auto_tab_active.py", press=["space"])
+
+
+def test_hatch(snap_compare):
+    """Test hatch styles."""
+    assert snap_compare(SNAPSHOT_APPS_DIR / "hatch.py")
+
+
+def test_rules(snap_compare):
+    """Test rules."""
+    assert snap_compare(SNAPSHOT_APPS_DIR / "rules.py")